+++ nflow-engine/src/main/java/io/nflow/engine/workflow/definition/StateExecution.java
@@ -35,7 +35,7 @@ public interface StateExecution {
    */
   int getRetries();
 
-  /**
+    /**
    * Return a string value of the given variable.
    *
    * @param name The name of the variable.