+++ nflow-tests/src/test/java/io/nflow/tests/AbstractNflowTest.java
@@ -2,6 +2,7 @@ package io.nflow.tests;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_EMPTY;
 import static java.lang.Thread.sleep;
+import static javax.ws.rs.core.MediaType.APPLICATION_JSON_TYPE;
 import static org.apache.cxf.jaxrs.client.WebClient.fromClient;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -15,6 +16,7 @@ import javax.inject.Named;
 import javax.ws.rs.core.UriBuilder;
 
 import org.apache.cxf.jaxrs.client.WebClient;
+import org.joda.time.Period;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
@@ -27,6 +29,9 @@ import io.nflow.rest.v1.msg.CreateWorkflowInstanceRequest;
 import io.nflow.rest.v1.msg.CreateWorkflowInstanceResponse;
 import io.nflow.rest.v1.msg.ListWorkflowDefinitionResponse;
 import io.nflow.rest.v1.msg.ListWorkflowInstanceResponse;
+import io.nflow.rest.v1.msg.MaintenanceRequest;
+import io.nflow.rest.v1.msg.MaintenanceRequest.MaintenanceRequestItem;
+import io.nflow.rest.v1.msg.MaintenanceResponse;
 import io.nflow.rest.v1.msg.SetSignalRequest;
 import io.nflow.rest.v1.msg.StatisticsResponse;
 import io.nflow.rest.v1.msg.UpdateWorkflowInstanceRequest;
@@ -36,9 +41,10 @@ import io.nflow.rest.v1.msg.WorkflowDefinitionStatisticsResponse;
 import io.nflow.tests.config.PropertiesConfiguration;
 import io.nflow.tests.config.RestClientConfiguration;
 import io.nflow.tests.extension.NflowServerConfig;
+import io.nflow.tests.extension.NflowServerExtension;
 import io.nflow.tests.extension.SkipTestMethodsAfterFirstFailureExtension;
 
-@ExtendWith({ SpringExtension.class, SkipTestMethodsAfterFirstFailureExtension.class })
+@ExtendWith({ NflowServerExtension.class, SpringExtension.class, SkipTestMethodsAfterFirstFailureExtension.class })
 @ContextConfiguration(classes = { RestClientConfiguration.class, PropertiesConfiguration.class })
 public abstract class AbstractNflowTest {
   protected WebClient workflowInstanceResource;
@@ -173,6 +179,15 @@ public abstract class AbstractNflowTest {
     mapper.setSerializationInclusion(NON_EMPTY);
     mapper.registerModule(new JodaModule());
     return mapper;
+  }
+
+  protected void deleteAllFinishedWorkflows() {
+    MaintenanceRequest req = new MaintenanceRequest();
+    req.deleteWorkflows = new MaintenanceRequestItem();
+    req.deleteWorkflows.olderThanPeriod = Period.seconds(0);
+    req.deleteArchivedWorkflows = req.deleteWorkflows;
+    assertTimeoutPreemptively(Duration.ofSeconds(15),
+        () -> fromClient(maintenanceResource).type(APPLICATION_JSON_TYPE).post(req, MaintenanceResponse.class));
   }
 
   protected String updateWorkflowInstance(long instanceId, UpdateWorkflowInstanceRequest request) {