+++ nflow-tests/src/test/java/io/nflow/tests/MaintenanceTest.java
@@ -21,7 +21,6 @@ import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.nflow.rest.v1.msg.CreateWorkflowInstanceRequest;
 import io.nflow.rest.v1.msg.CreateWorkflowInstanceResponse;
@@ -30,9 +29,7 @@ import io.nflow.rest.v1.msg.MaintenanceRequest.MaintenanceRequestItem;
 import io.nflow.rest.v1.msg.MaintenanceResponse;
 import io.nflow.tests.demo.workflow.FibonacciWorkflow;
 import io.nflow.tests.extension.NflowServerConfig;
-import io.nflow.tests.extension.NflowServerExtension;
 
-@ExtendWith(NflowServerExtension.class)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 public class MaintenanceTest extends AbstractNflowTest {
   private static final int STEP_1_WORKFLOWS = 4;
@@ -51,8 +48,7 @@ public class MaintenanceTest extends AbstractNflowTest {
   @Test
   @Order(1)
   public void cleanupExistingStuff() {
-    archiveOlderThan(now());
-    deleteOlderThan(now());
+    deleteAllFinishedWorkflows();
   }
 
   @Test